Write method for model summary
Move the code from the Predictions class to Model. Add other methods to
get and save the model and training parameters. Training history now
keep all information, which can be retrieved and save with a dedicated
method (also called in summary). Average of history is performed on demand.

Fix predictions of NeuralNet (broken for sequential models).