Add learning curve and metric lookup
Learning curves can be computed from Model. A plot function has been
added in the module describe.

To help in evaluating the performance for learning curves, a method
evaluate has been added to Model. The latter uses a new lookup mechanism
for metrics. The latter can discoved the metrics compatible with the
data at hand, and also find the best naming depending on the context.